+++ weather-you/app/build.gradle
@@ -96,10 +96,10 @@ static def computeVersionName() {
 }
 
 static def computeVersionCode() {
-    return (Sdk.phoneVersionMajor * 100000)
+    return ((Sdk.phoneVersionMajor * 100000)
     + (Sdk.phoneVersionMinor * 10000)
     + (Sdk.phoneVersionPatch * 1000)
-    + Integer.valueOf(System.getenv("CIRCLE_BUILD_NUM") ?: 0)
+    + Integer.valueOf(System.getenv("CIRCLE_BUILD_NUM") ?: Sdk.localVersionCode))
 }
 
 dependencies {
