+++ weather-you/buildSrc/src/main/java/Sdk.kt
@@ -15,8 +15,10 @@ object Sdk {
 
     // TV
     const val tvVersionMajor = 3
-    const val tvVersionMinor = 0
-    const val tvVersionPatch = 1
+    const val tvVersionMinor = 1
+    const val tvVersionPatch = 0
+
+    const val localVersionCode = 41
 
     const val applicationId = "com.rodrigmatrix.weatheryou"
 }
