+++ weather-you/common/components/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="app_name" translatable="false">WeatherYou</string>
     <string name="feels_like_x">Feels Like %s</string>
-    <string name="high_lowercase_x" translatable="false">H%s</string>
-    <string name="low_lowercase_x" translatable="false">L%s</string>
+    <string name="high_lowercase_x">H%s</string>
+    <string name="low_lowercase_x">L%s</string>
     <string name="day_x">Day %s</string>
     <string name="night_x">Night %s</string>
     <string name="next_x_days_forecast">Next %d Days Forecast</string>
@@ -76,7 +76,7 @@
     <string name="minutes_x">%dm</string>
     <string name="current_weather_small_widget_description">Displays the current weather</string>
     <string name="no_results_found">No results found</string>
-    <string name="keep_typing">Keep typing...</string>
+    <string name="keep_typing">Keep typing…</string>
     <string name="location">Location</string>
     <string name="locations">Locations</string>
     <string name="units">Units</string>
@@ -133,11 +133,12 @@
     <string name="enable_weather_animations">Enable Weather Animations</string>
     <string name="click_to_search">Click on the search icon to search</string>
 
-    <string name="updating_location">Updating location...</string>
+    <string name="updating_location">Updating location…</string>
 
     <string name="apple_weather">Apple Weather</string>
     <string name="for_more_information">For more information, </string>
     <string name="visit_apple_weather">visit Apple Weather</string>
+    <string name="actual">Actual</string>
     <string name="feels_like">Feels Like</string>
     <string name="celsius">Celsius (°C)</string>
     <string name="fahrenheit">Fahrenheit (°F)</string>
@@ -147,6 +148,7 @@
 
     <string name="day_chance">%s\'s chance: %s</string>
     <string name="actual_temperature">Actual: %s</string>
+
     <string name="millimeters">mm</string>
     <string name="centimeters">cm</string>
 </resources>
