+++ weather-you/features/addlocation/build.gradle
@@ -68,6 +68,7 @@ dependencies {
 
     implementation platform(libs.firebase.bom)
     implementation(libs.firebase.crashlytics)
+    implementation(libs.firebase.analytics)
 
     implementation(libs.jodaTime)
