+++ weather-you/features/addlocation/src/main/java/com/rodrigmatrix/weatheryou/addlocation/di/AddLocationModule.kt
@@ -19,6 +19,7 @@ object AddLocationModule {
                 getFamousLocationsUseCase = get(),
                 searchLocationUseCase = get(),
                 adsManager = get(),
+                firebaseAnalytics = get(),
                 firebaseCrashlytics = FirebaseCrashlytics.getInstance(),
             )
         }
