+++ weather-you/features/locationdetails/src/main/java/com/rodrigmatrix/weatheryou/locationdetails/presentaion/conditions/ConditionsViewState.kt
@@ -7,8 +7,20 @@ data class ConditionsViewState(
     val weatherLocation: WeatherLocation? = null,
     val day: WeatherDay? = null,
     val isCurrentDay: Boolean = false,
+    val type: ConditionType = ConditionType.Conditions,
     val temperatureType: TemperatureType = TemperatureType.Actual,
 )
+
+enum class ConditionType {
+    Conditions,
+    SunriseSunset,
+//    UvIndex,
+//    Wind,
+//    Precipitation,
+//    Humidity,
+//    Visibility,
+//    Pressure,
+}
 
 enum class TemperatureType {
     Actual,
