+++ weather-you/features/locationdetails/src/main/java/com/rodrigmatrix/weatheryou/locationdetails/presentaion/details/WeatherDetailsScreen.kt
@@ -359,7 +359,7 @@ fun AppleWeatherAttribution(
                     }
                 )
         )
-        Spacer(Modifier.height(16.dp))
+        Spacer(Modifier.height(32.dp))
     }
 }
